restructure: moved otel module from handlers to src
