+++ src/otel.rs
@@ -16,36 +16,7 @@
  *
  */
 
-mod about;
-mod alerts;
-pub mod analytics;
-pub mod banner;
-mod catalog;
-mod cli;
-pub mod correlation;
-mod event;
-pub mod handlers;
-pub mod hottier;
-pub mod kafka;
-mod livetail;
-mod metadata;
+pub mod logs;
 pub mod metrics;
-pub mod migration;
-mod oidc;
-pub mod option;
-pub mod otel;
-mod query;
-pub mod rbac;
-mod response;
-mod static_schema;
-mod stats;
-pub mod storage;
-pub mod sync;
-pub mod users;
-mod utils;
-mod validator;
-pub use handlers::http::modal::{
-    ingest_server::IngestServer, query_server::QueryServer, server::Server, ParseableServer,
-};
-
-pub const STORAGE_UPLOAD_INTERVAL: u32 = 60;
+pub mod otel_utils;
+pub mod traces;
